pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "Hi-C"
version = "0.0.1"
authors = [
    { name="Robert Loftin", email="R.T.Loftin@tudelft.nl"},
]
description = 'Experiments for the paper "Uncoupled Learning of Differential Stackelberg Equilibria with Commitments" (https://arxiv.org/abs/2302.03438)'
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "GitPython==3.1.31",
    "numpy==1.24.0",
    "pandas==1.5.3",
    "PyYAML==6.0",
    "tensorboardX==2.6"
]

[project.optional-dependencies]
torch = ["torch==1.13.1"]

[project.urls]
"Homepage" = "https://github.com/rtloftin/Hi-C"
"Bug Tracker" = "https://github.com/rtloftin/Hi-C/issues"